Fixed and added code for the edn stats. add some bounce for the walls

+++ Assets/Scripts/UpdateStats.cs
@@ -142,10 +142,10 @@ public class UpdateStats : MonoBehaviour {
             strikesText.color = Color.red;
         }
     }
-    public int GetNumPitches()
+    public float GetNumPitches()
     {
         int numPitches = strikes + hits;
-        return numPitches;
+        return (float)numPitches;
     }
 }
